% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dix_print.R
\name{dix_print}
\alias{dix_print}
\title{dix_print}
\usage{
dix_print(print_in01 = NA, print_in02 = NA, print_in03 = NA)
}
\value{
string
}
\description{
text generation function from transmitted parameters
}
\details{
dix
}
\examples{
dix_print

}
